refactorizacion de los ejercicios
